pyproject.toml
[tool.poetry]
name = "sc-20230710-sce"
version = "0.1.0"
description = ""
authors = ["SongshGeo <songshgeo@gmail.com>"]
readme = "README.md"
packages = [{include = "sc_20230710_sce"}]

[tool.poetry.dependencies]
python = "^3.11"
hydra-core = "^1.3.2"
abses = {git="https://github.com/SongshGeo/ABSESpy.git", branch='dev'}
mesa-geo = ">=0.5.0"
fiona = "^1.9.4.post1"
rasterio = "^1.3.8"
seaborn = "^0.12.2"
rioxarray = "^0.15.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
allure-pytest = "^2.13.2"
ipykernel = "^6.25.1"
jupyterlab = "^4.0.5"
jupyterlab-execute-time = "^3.0.1"
flake8 = "^6.1.0"
isort = "^5.12.0"
nbstripout = "^0.6.1"
pydocstyle = "^6.3.0"
pre-commit-hooks = "^4.4.0"
sourcery = "^1.8.0"
pytest-cov = "^4.1.0"
pytest-clarity = "^1.0.1"
pytest-sugar = "^0.9.7"
interrogate = "^1.5.0"
mypy = "^1.5.1"
bandit = "^1.7.5"
black = "^23.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
